+++ tdmclient/__main__.py
@@ -14,6 +14,7 @@ import sys
 import getopt
 
 tool_set = {
+    "info",
     "gui",
     "list",
     "repl",